+++ src/main/java/com/notenoughmail/kubejs_tfc/util/implementation/ItemStackProviderJS.java
@@ -66,6 +66,13 @@ public record ItemStackProviderJS(ItemStack stack, JsonArray modifiers) implemen
         return new ItemStackProviderJS(ItemStackJS.of(o), new JsonArray());
     }
 
+    public static ItemStackProvider ofCannon(@Nullable Object o) {
+        if (o instanceof ItemStackProvider isp) {
+            return isp; // Skip to serialize & deserialize trip when canonizing
+        }
+        return of(o).asCanonClass();
+    }
+
     @Override
     public Object replaceOutput(RecipeJS recipe, ReplacementMatch match, OutputReplacement original) {
         if (recipe instanceof ISupportProviderOutput) {